Thing angles outside of 0-359 are now preserved when copypasting or placing new Things

+++ Source/Core/Config/ProgramConfiguration.cs
@@ -141,7 +141,7 @@ namespace CodeImp.DoomBuilder.Config
 		private int defaultfloorheight;
 		private int defaultceilheight;
 		private int defaultthingtype = 1;
-		private float defaultthingangle;
+		private int defaultthingangle;
 		private List<string> defaultthingflags;
 		
 		#endregion
@@ -251,7 +251,7 @@ namespace CodeImp.DoomBuilder.Config
 		public int DefaultCeilingHeight { get { return defaultceilheight; } set { defaultceilheight = value; } }
 
 		public int DefaultThingType { get { return defaultthingtype; } set { defaultthingtype = value; } }
-		public float DefaultThingAngle { get { return defaultthingangle; } set { defaultthingangle = value; } }
+		public int DefaultThingAngle { get { return defaultthingangle; } set { defaultthingangle = value; } }
 
 		#endregion
